app/requirements.txt
black==19.10b0
Django==3.0.5
dj-database-url==0.5.0
djangorestframework==3.11.0
flake8==3.7.9
gunicorn==20.0.4
isort==4.3.21
psycopg2-binary==2.8.5
pytest-cov==2.8.1
pytest-django==3.9.0
pytest==5.4.1
whitenoise==5.0.1